+++ app/Utils/Backup.php
@@ -1,26 +1,21 @@
 <?php
-
 /**
- * Settings Backups module model class.
+ * Backup class.
+ *
+ * @package   App
  *
  * @copyright YetiForce Sp. z o.o
  * @license   YetiForce Public License 3.0 (licenses/LicenseEN.txt or yetiforce.com)
  * @author    Arkadiusz Dudek <a.dudek@yetiforce.com>
  */
-class Settings_Backups_Module_Model extends Settings_Vtiger_Module_Model
+
+namespace App\Utils;
+
+/**
+ * Backup.
+ */
+class Backup
 {
-	public $name = 'Backups';
-
-	/**
-	 * Return catalog with backup files.
-	 *
-	 * @return string
-	 */
-	public static function getCatalogPath()
-	{
-		return \AppConfig::module('Backups', 'BACKUPS_PATH');
-	}
-
 	/**
 	 * Read catalog with backup files and return catalogs and files list.
 	 *
@@ -30,15 +25,15 @@ class Settings_Backups_Module_Model extends Settings_Vtiger_Module_Model
 	 *
 	 * @return array
 	 */
-	public static function readCatalog(\App\Request $request)
+	public static function readCatalog($catalogToRead, $module)
 	{
-		$catalogToRead = $request->getByType('catalog', 'String');
-		$catalogPath = self::getCatalogPath();
+		//$catalogToRead = $request->getByType('catalog', 'String');
+		$catalogPath = self::getBackupCatalogPath();
 		$catalogToReadArray = $returnStructure = [];
 		$urlDirectory = '';
 		if (!empty($catalogToRead)) {
 			$catalogToReadArray = explode(DIRECTORY_SEPARATOR, $catalogToRead);
-			$catalogPath = self::getCatalogPath() . DIRECTORY_SEPARATOR . $catalogToRead;
+			$catalogPath .= DIRECTORY_SEPARATOR . $catalogToRead;
 			$urlDirectory = $catalogToRead . DIRECTORY_SEPARATOR;
 		}
 		if (!self::isAllowedDirectory($catalogToRead)) {
@@ -46,7 +41,7 @@ class Settings_Backups_Module_Model extends Settings_Vtiger_Module_Model
 		}
 		$catalogs = array_diff(scandir($catalogPath, SCANDIR_SORT_ASCENDING), ['.']);
 		foreach ($catalogs as $element) {
-			$requestUrl = 'index.php?module=' . $request->getModule() . '&parent=Settings&view=Index';
+			$requestUrl = "index.php?module=$module&parent=Settings&view=Index";
 			if ('..' === $element) {
 				if (!empty($catalogToReadArray)) {
 					array_pop($catalogToReadArray);
@@ -69,6 +64,20 @@ class Settings_Backups_Module_Model extends Settings_Vtiger_Module_Model
 	}
 
 	/**
+	 * Return catalog with backup files.
+	 *
+	 * @return string
+	 */
+	public static function getBackupCatalogPath()
+	{
+		$backupPath = \AppConfig::module('Backup', 'BACKUP_PATH');
+		if (empty($backupPath)) {
+			throw new \App\Exceptions\NoPermittedForAdmin(\App\Language::translate('LBL_CONFIGURE_BEFORE_USE'));
+		}
+		return $backupPath;
+	}
+
+	/**
 	 * Check is it an allowed directory.
 	 *
 	 * @param $dir
@@ -78,7 +87,7 @@ class Settings_Backups_Module_Model extends Settings_Vtiger_Module_Model
 	public static function isAllowedDirectory($dir)
 	{
 		$isAllowed = true;
-		$fullPath = self::getCatalogPath() . DIRECTORY_SEPARATOR . $dir;
+		$fullPath = self::getBackupCatalogPath() . DIRECTORY_SEPARATOR . $dir;
 		if (!is_writable($fullPath) || !is_dir($fullPath) || is_file($fullPath) || strpos($fullPath, '../') !== false || strpos($fullPath, '..\\') !== false) {
 			$isAllowed = false;
 		}
@@ -95,7 +104,7 @@ class Settings_Backups_Module_Model extends Settings_Vtiger_Module_Model
 	public static function isAllowedFileDirectory($dir)
 	{
 		$isAllowed = true;
-		$fullPath = self::getCatalogPath() . DIRECTORY_SEPARATOR . $dir;
+		$fullPath = self::getBackupCatalogPath() . DIRECTORY_SEPARATOR . $dir;
 		if (!is_writable($fullPath) || is_dir($fullPath) || !is_file($fullPath) || strpos($fullPath, '../') !== false || strpos($fullPath, '..\\') !== false) {
 			$isAllowed = false;
 		}
